+++ app/pages/sections/catalogue-solutions/order-item/flat/ondemand/ui-tests/general.ui.test.js
@@ -99,70 +99,6 @@ fixture('Catalogue-solutions - recipient page - general')
     await t.expect(isDone).ok('Not all nock interceptors were used!');
   });
 
-test('should render legend with mainAdvice', async (t) => {
-  await pageSetup();
-  await t.navigateTo(pageUrl);
-
-  const mainAdvice = Selector('legend');
-
-  await t
-    .expect(mainAdvice.exists).ok()
-    .expect(await extractInnerText(mainAdvice)).eql(content.questions.plannedDeliveryDate.mainAdvice);
-});
-
-test('should render additionalAdvice', async (t) => {
-  await pageSetup();
-  await t.navigateTo(pageUrl);
-
-  const additionalAdvice = Selector('[data-test-id="date-field-input"] span.nhsuk-hint');
-
-  await t
-    .expect(additionalAdvice.exists).ok()
-    .expect(await extractInnerText(additionalAdvice)).eql(content.questions.plannedDeliveryDate.additionalAdvice);
-});
-
-test('should render labels for day, month and year inputs', async (t) => {
-  await pageSetup();
-  await t.navigateTo(pageUrl);
-
-  const labels = Selector('label');
-  const dayLabel = labels.nth(0);
-  const monthLabel = labels.nth(1);
-  const yearLabel = labels.nth(2);
-
-  await t
-    .expect(dayLabel.exists).ok()
-    .expect(await extractInnerText(dayLabel)).eql('Day')
-    .expect(monthLabel.exists).ok()
-    .expect(await extractInnerText(monthLabel)).eql('Month')
-    .expect(yearLabel.exists).ok()
-    .expect(await extractInnerText(yearLabel)).eql('Year');
-});
-
-test('should render input fields for day, month and year', async (t) => {
-  await pageSetup();
-  await t.navigateTo(pageUrl);
-
-  const inputFields = Selector('#plannedDeliveryDate input:not([name=_csrf])');
-  const dayInput = inputFields.nth(0);
-  const monthInput = inputFields.nth(1);
-  const yearInput = inputFields.nth(2);
-
-  await t
-    .expect(dayInput.exists).ok()
-    .expect(dayInput.getAttribute('id')).eql('plannedDeliveryDate-day')
-    .expect(dayInput.getAttribute('name')).eql('plannedDeliveryDate-day')
-    .expect(dayInput.getAttribute('type')).eql('number')
-    .expect(monthInput.exists).ok()
-    .expect(monthInput.getAttribute('id')).eql('plannedDeliveryDate-month')
-    .expect(monthInput.getAttribute('name')).eql('plannedDeliveryDate-month')
-    .expect(monthInput.getAttribute('type')).eql('number')
-    .expect(yearInput.exists).ok()
-    .expect(yearInput.getAttribute('id')).eql('plannedDeliveryDate-year')
-    .expect(yearInput.getAttribute('name')).eql('plannedDeliveryDate-year')
-    .expect(yearInput.getAttribute('type')).eql('number');
-});
-
 test('should render a text field for the quantity question', async (t) => {
   await pageSetup();
   await t.navigateTo(pageUrl);
